requirements.txt
llama-cpp-python>=0.3.12
aiofiles>=23.0.0
